requirements.txt
fastapi
uvicorn
pymongo
faiss-cpu
redis
celery
deep-translator
python-dotenv
requests
groq
beautifulsoup4
langchain
youtube-search-python
